+++ app/src/main/java/com/jorgesm/themoviedb/data/datasource/LocationDataSource.kt
@@ -1,0 +1,9 @@
+package com.jorgesm.themoviedb.data.datasource
+
+
+import android.location.Location
+
+interface LocationDataSource {
+    suspend fun findLastLocation(): Location?
+}
+
