+++ app/src/main/java/com/jorgesm/themoviedb/framework/database/MovieDao.kt
@@ -1,4 +1,4 @@
-package com.jorgesm.themoviedb.data.database
+package com.jorgesm.themoviedb.framework.database
 
 import androidx.room.Dao
 import androidx.room.Insert
